DNIe user interface moved into card-dnie.c.
