+++ src/libopensc/Makefile.mak
@@ -23,7 +23,7 @@ OBJECTS			= \
 	card-rtecp.obj card-westcos.obj card-myeid.obj card-ias.obj \
 	card-itacns.obj card-authentic.obj \
 	card-iasecc.obj iasecc-sdo.obj iasecc-sm.obj cwa-dnie.obj cwa14890.obj \
-	card-sc-hsm.obj card-dnie.obj user-interface.obj card-isoApplet.obj \
+	card-sc-hsm.obj card-dnie.obj card-isoApplet.obj \
 	card-masktech.obj \
 	\
 	pkcs15-openpgp.obj pkcs15-infocamere.obj pkcs15-starcert.obj \
